UseSession a medias (le toca G auyda)

+++ src/app/Dashboard/page.jsx
@@ -6,17 +6,31 @@ import axios from "axios";
 import { useEffect, useState } from "react";
 import Modal from "./Modal";
 import Modal2 from "./Modal2";
+import { useSession } from "next-auth/react";
 
 export default function Home() {
+
+  const { data: session, status } = useSession([]);
+  console.log(data, status)
   const [teams, setTeams] = useState([]);
   const [startIndex, setStartIndex] = useState(0);
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [isModalOpen2, setIsModalOpen2] = useState(false);
   const [isModalOpen3, setIsModalOpen3] = useState(false);
   const [selectedTeamId, setSelectedTeamId] = useState(null);
+
+  useEffect(() => {
+    if (status === "authenticated") {
+      fetchUserTeams();
+    }
+  }, [status]);
+
   const fetchUserTeams = async () => {
     try {
-      const userId = 36;
+      const userId = session?.user?.id;
+      if (!userId) {
+        return;
+      }
       const response = await axios.get(
         `http://localhost:3005/api/teams/${userId}/teams`
       );
@@ -40,7 +54,7 @@ export default function Home() {
   };
 
   const openModal = (teamId) => {
-    console.log("Opening modal for directory ID:", teamId); 
+    console.log("Opening modal for directory ID:", teamId);
     setSelectedTeamId(teamId);
     setIsModalOpen3(true);
   };
@@ -51,16 +65,26 @@ export default function Home() {
 
   const handleDelete = async () => {
     try {
-      const userId=36;
+      const userId = session?.user?.id;
+      if (!userId) {
+        return;
+      }
+
       console.log("Deleting directory ID:", selectedTeamId);
       const url = `http://localhost:3005/api/teams/user-teams/${userId}/${selectedTeamId}`;
       await axios.delete(url);
-      setTeams(teams => teams.filter(t => t.teamId !== selectedTeamId));
+      setTeams((teams) => teams.filter((t) => t.teamId !== selectedTeamId));
       closeModal();
     } catch (error) {
-      console.error('Error deleting directory:', error);
+      console.error("Error deleting directory:", error);
     }
   };
+  if (status === "loading") {
+    return <div>Loading...</div>;
+  }
+  if (status === "unauthenticated") {
+    return <div>Unauthenticated</div>;
+  }
 
   return (
     <main className="flex flex-col min-h-screen">
@@ -104,7 +128,10 @@ export default function Home() {
                 style={{ width: 363, height: 240 }}
               >
                 <div className="table-r-r self-end">
-                      <i className="bi bi-trash3-fill text-gray-300 text-xl " onClick={() => openModal(team.teamId)}></i>
+                  <i
+                    className="bi bi-trash3-fill text-gray-300 text-xl "
+                    onClick={() => openModal(team.teamId)}
+                  ></i>
                 </div>
                 <img
                   style={{ marginBottom: "10px" }}
@@ -144,15 +171,26 @@ export default function Home() {
       {isModalOpen3 && (
         <div className="modal">
           <div className="modal-content ">
-            <h2 className="font-bold text-xl" style={{ color: '#24374B' }} >You want to delete this team?</h2>
+            <h2 className="font-bold text-xl" style={{ color: "#24374B" }}>
+              You want to delete this team?
+            </h2>
             <div className="button-container">
-              <button className="bg-[#E92525] text-white p-2 rounded-lg mt-5 px-10" onClick={handleDelete}>Delete</button>
-              <button className="bg-[#768396] text-white p-2 rounded-lg mt-5 px-10" onClick={closeModal}>Cancel</button>
+              <button
+                className="bg-[#E92525] text-white p-2 rounded-lg mt-5 px-10"
+                onClick={handleDelete}
+              >
+                Delete
+              </button>
+              <button
+                className="bg-[#768396] text-white p-2 rounded-lg mt-5 px-10"
+                onClick={closeModal}
+              >
+                Cancel
+              </button>
             </div>
           </div>
         </div>
       )}
-
     </main>
   );
 }
